# docs/swagger.yaml
basePath: /v1
definitions:
...
  main.RegisterMentorPayload:
    properties:
      country:
        type: string
      language:
        items:
          type: string
        type: array
      name:
        maxLength: 40
        type: string
    required:
    - country
    - language
    - name
    type: object
...
  main.UpdateMentorPayload:
    properties:
      country:
        type: string
      language:
        items:
          type: string
        type: array
      name:
        maxLength: 40
        type: string
    type: object
...
  store.BookingSlot:
    properties:
      created_at:
        type: string
      days:
        items:
          type: string
        type: array
      end_period:
        type: string
      end_time:
        type: string
      id:
        type: integer
      start_period:
        type: string
      start_time:
        type: string
      updated_at:
        type: string
      userid:
        type: integer
    type: object
...
  store.Education:
    properties:
      degree:
        type: string
      field:
        type: string
      id:
        type: integer
      institute:
        type: string
      userid:
        type: integer
      year_from:
        type: string
      year_to:
        type: string
    type: object
  store.Experience:
    properties:
      company:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      userid:
        type: integer
      year_from:
        type: string
      year_to:
        type: string
    type: object
...
  store.Gig:
    properties:
      created_at:
        type: string
      description:
        type: string
      discipline:
        items:
          type: string
        type: array
      expertise:
        type: string
      id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      userid:
        type: integer
    type: object
  store.Mentor:
    properties:
      bookingslots:
        items:
          $ref: '#/definitions/store.BookingSlot'
        type: array
      country:
        type: string
      created_at:
        type: string
      education:
        items:
          $ref: '#/definitions/store.Education'
        type: array
      experience:
        items:
          $ref: '#/definitions/store.Experience'
        type: array
      gigs:
        items:
          $ref: '#/definitions/store.Gig'
        type: array
      id:
        type: integer
      language:
        items:
          type: string
        type: array
      name:
        type: string
      updated_at:
        type: string
      userid:
        type: integer
      workingat:
        $ref: '#/definitions/store.WorkingAt'
    type: object
...
  store.WorkingAt:
    properties:
      company:
        type: string
      github:
        type: string
      id:
        type: integer
      instagram:
        type: string
      linkedin:
        type: string
      month:
        type: integer
      title:
        type: string
      totalyear:
        type: integer
      userid:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a Appointr server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Appointr API
  version: "1.0"
paths:
...
  /mentors:
    get:
      consumes:
      - application/json
      description: Get all mentors
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.Mentor'
            type: array
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get all mentors
      tags:
      - mentor
  /mentors/{mentorID}:
    delete:
      consumes:
      - application/json
      description: Delete mentor
      parameters:
      - description: Mentor ID
        in: path
        name: mentorID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Delete mentor
      tags:
      - mentor
    get:
      consumes:
      - application/json
      description: Get mentor by ID
      parameters:
      - description: Mentor ID
        in: path
        name: mentorID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Mentor'
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get mentor by ID
      tags:
      - mentor
    patch:
      consumes:
      - application/json
      description: Update mentor
      parameters:
      - description: Mentor ID
        in: path
        name: mentorID
        required: true
        type: integer
      - description: Mentor payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.UpdateMentorPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Mentor'
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Update mentor
      tags:
      - mentor
  /mentors/create:
    post:
      consumes:
      - application/json
      description: Creates a mentor
      parameters:
      - description: Mentor payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.RegisterMentorPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.Mentor'
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Creates a mentor
      tags:
      - mentor
  /mentors/dis/{discipline}:
    get:
      consumes:
      - application/json
      description: Get mentor by discipline
      parameters:
      - description: Expertise
        in: path
        name: discipline
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.Mentor'
            type: array
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get mentor by discipline
      tags:
      - mentor
  /mentors/exp/{expertise}:
    get:
      consumes:
      - application/json
      description: Get mentor by expertise
      parameters:
      - description: Expertise
        in: path
        name: expertise
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.Mentor'
            type: array
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get mentor by expertise
      tags:
      - mentor
  /mentors/name/{mentorName}:
    get:
      consumes:
      - application/json
      description: Get mentor by name
      parameters:
      - description: Mentor Name
        in: path
        name: mentorName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.Mentor'
            type: array
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get mentor by name
      tags:
      - mentor
  /mentors/u/{userID}:
    get:
      consumes:
      - application/json
      description: Get mentor by userID
      parameters:
      - description: Mentor ID
        in: path
        name: userID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Mentor'
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get mentor by userID
      tags:
      - mentor
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"